main features finished and classloaderWIP

+++ src/ck/ckrc/erie/warchess/game/DamageListener.java
@@ -2,6 +2,6 @@ package ck.ckrc.erie.warchess.game;
 
 public interface DamageListener {
 
-    public int takeDamage(int damage);
+    public double takeDamage(double damage);
 
 }
